+++ progc/header.h
@@ -10,14 +10,14 @@
         struct arbre* gauche;
         struct arbre* droite;
         int eq;
-        int elmnt;
+        float elmnt;
         char* csv; //contient toutes les autres infos de l'élément (qui ne servent pas au tri)
     }Arbre;
 
     int min(int a, int b);
     int max(int a, int b);
-    Arbre* creerarbre(int x, char* c);
-    Arbre* ajoutabr(Arbre* a, int x, char* c, int* h);
+    Arbre* creerarbre(float x, char* c);
+    Arbre* ajoutabr(Arbre* a, float x, char* c, int* h);
     Arbre* equilibreravl(Arbre* a);
     void traitement_t(int argc, char* argv[]);
     void traitement_s(int argc, char* argv[]);
